refactor : dto가 필요한 곳에서만 활용할 수 있도록 default 접근제한자로 수정하고 패키지 수정
