+++ src/main/java/chess/mysql/chess/ChessConnection.java
@@ -1,4 +1,4 @@
-package chess.mysql;
+package chess.mysql.chess;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
