+++ src/main/java/chess/mysql/chess/ChessDao.java
@@ -1,4 +1,4 @@
-package chess.mysql;
+package chess.mysql.chess;
 
 import java.util.List;
 import java.util.Optional;
