+++ src/test/java/chess/controller/web/RoomControllerTest.java
@@ -1,5 +1,6 @@
 package chess.controller.web;
 
+import chess.controller.RoomRequestDto;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
