+++ src/test/java/chess/mysql/dao/ChessConnectionTest.java
@@ -1,6 +1,6 @@
 package chess.mysql.dao;
 
-import chess.mysql.ChessConnection;
+import chess.mysql.chess.ChessConnection;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
